Squashed commit of the following:
commit 56d4b38816526e87101b3c0de4eb548047af0dd0
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Wed Feb 16 12:22:37 2022 +0900

    big restructuring

commit 100f306e9322ffdbb23e23960ce8b2805391a6b5
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Mon Feb 14 18:13:51 2022 +0900

    converted

commit aa0ebf48496f04cdf08abd17b9dea2375007bed3
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Mon Feb 14 18:13:46 2022 +0900

    Overwrite

commit d7c87d6be3f02e3cb146e2fc1449aec898841ae8
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Mon Feb 14 17:32:10 2022 +0900

    Convert

commit a55a8a7da09dc73ed18301358bb8a789af4f6840
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Mon Feb 14 13:51:10 2022 +0900

    from ulong to uint

commit 958ea017b202bc72554d062c828c55c8b6e3795d
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Mon Feb 14 13:46:07 2022 +0900

    temporary group the network response types

commit 35e4279b5b0520ffd815d7411ada5927136a58e0
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Mon Feb 14 13:45:52 2022 +0900

    remove messagepacks

commit fca5bb1c2b8130058d21932fd5f84535ad9f0cc9
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Mon Feb 14 12:45:51 2022 +0900

    shutdown 1 day

commit 9fb6a30252eda0337b8d4fe4e55484e272af355b
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 22:23:21 2022 +0900

    namespace PixivApi to PixivApi.Console

commit c4a7727d070659e781abce7efd2fd08935aaab32
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 17:34:01 2022 +0900

    remove unused

commit 4cd2cd6d5d58f7e503387d4d37c7930e8d58bd76
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 17:14:14 2022 +0900

    merge prefix support

commit f196886764eeb94550a67724c52ae3d39d41c15d
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 17:13:53 2022 +0900

    update

commit b4ed6ad5b6a3aa651aab29de351a467f11be233d
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 16:01:26 2022 +0900

    gitignore

commit 1a80024a3e67db10885fa4116e3e769b108a81ee
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 16:01:21 2022 +0900

    Make Bidirectional set

commit c1f2de99ef65dc5512b66e507791c906e4e74357
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 16:01:10 2022 +0900

    Hide

commit abc369460fe07daf2fe10105887d0af284c27987
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 16:01:03 2022 +0900

    Remove unused

commit fe2d4f5226a221bc8e1c45f83651ee5a43a8addc
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 16:00:54 2022 +0900

    remove usings

commit da6a85708370a2bba02a35bafe6bcd36330efa89
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 15:17:04 2022 +0900

    Usings

commit 653a03f91ed764f15ea2cb042326eac026e4781d
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 14:04:03 2022 +0900

    Add UgoiraFilter

commit c3462aa6905bd5a167b6bd136b5ecca4f2cb290f
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 13:42:20 2022 +0900

    not parallel

commit 0266906d69789859e17b64a885c243f512999907
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 13:28:56 2022 +0900

    UgoiraMetadata

commit 612a5cb211154c9204c7e9cd86d8a01cb2aa93c4
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 13:28:49 2022 +0900

    Add global using

commit 3d958c0821213a8b8cf7a3a711f71c9ed46d8033
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 11:26:42 2022 +0900

    log handling

commit 57b6c9c76bf0d9a38a106cbf0e20f83425a6af0d
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 11:05:29 2022 +0900

    Remove empty file

commit a6aca26e50b594a9d9dce4f28296c806b5b68093
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 10:56:24 2022 +0900

    divide

commit da8cd3fd17740e24a6c51b4bc9814d008484d18c
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 10:51:36 2022 +0900

    filter null handling

commit f7b495f82baffe03a2cbee7bf33ef6372aeca056
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 10:51:22 2022 +0900

    remove unneccessary log

commit 7404454bf5bbec3de4981fb76720036142621cfc
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 10:42:02 2022 +0900

    Not Require Console.ReadLine

commit 520cd911d5ae35ccbef64a5e90cbf97bb774ecdc
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 10:00:10 2022 +0900

    Simplify filter json handling
    Add IJsonOnDeserialized

commit 9bfa936ddde9e643bebcbb38ad452cc89f346d93
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 09:44:02 2022 +0900

    Serialize Unordered Parallel way

commit 4b4a1385070e0ae27758220c93e5380eb687d3bc
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 00:20:33 2022 +0900

    Partition

commit 8097a79dfb37f8c3de5acca72958fb587b083a39
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 00:15:32 2022 +0900

    Add Microsoft.Toolkit.HighPerformance
    Chunked WriteFileAsync

commit 09e3e35f101bd21e41a606ff548bf2ea2318c514
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sun Feb 13 00:14:43 2022 +0900

    remove unnecessary dependency

commit 1ec41537cb45b11f4cd1c90c436b57bc99d2b472
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 21:48:11 2022 +0900

    WriteTo

commit 16305d28bd09de11dd84980dbfda9ed55609952d
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 21:48:04 2022 +0900

    Add using

commit a18b7fa3e5d670548f1d41af80c30092a3a63a83
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 21:47:56 2022 +0900

    Experimental

commit 2ffe1bd1630daa5d8abcd8cc0f67f41d93e932e0
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 20:22:24 2022 +0900

    make internal

commit d65ff1d8f96254b7b6e2cbe8d76ff823f9c256be
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 20:22:00 2022 +0900

    From record to struct

commit 445e01e122b0d09c5e67d149f735caab5e87aa80
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 20:21:49 2022 +0900

    static

commit e8b2e7355ff2a223a31d0aece8608b9533e9003c
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 20:21:04 2022 +0900

    options

commit 99fb6b8cf9741d139f55e2d97e8c189fe9830c44
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 20:20:11 2022 +0900

    Formattable

commit fd197c2c48e4120282a7a333d35ca2e2c015eb8d
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 15:49:48 2022 +0900

    pipe map

commit a0ed931523d8a2440e4e4823e4dce67034fc804e
Author: pCYSl5EDgo <pCYSl5EDgo@yahoo.co.jp>
Date:   Sat Feb 12 15:49:36 2022 +0900

    remove unfollow

+++ PixivApi.Console/Local/ClearZeroBytes.cs
@@ -1,15 +1,26 @@
-namespace PixivApi;
+namespace PixivApi.Console;
 
 partial class LocalClient
 {
     [Command("clear-0-bytes", "")]
     public ValueTask ClearZeroBytes(
         bool pipe = false
-    ) => ClearAsync(logger, configSettings, pipe, Context.CancellationToken);
+    ) => ClearAsync(logger, configSettings, pipe, false, Context.CancellationToken);
 
-    internal static async ValueTask ClearAsync(ILogger logger, ConfigSettings configSettings, bool pipe, CancellationToken token)
+    internal static async ValueTask ClearAsync(ILogger logger, ConfigSettings configSettings, bool pipe, bool forceClear, CancellationToken token)
     {
-        System.Collections.Concurrent.ConcurrentBag<FileInfo> files = new();
+        if (!pipe)
+        {
+            logger.LogInformation("Start clearing.");
+        }
+
+        if (forceClear)
+        {
+            await ForceClearAsync(logger, configSettings, pipe, token).ConfigureAwait(false);
+            return;
+        }
+
+        ConcurrentBag<FileInfo> files = new();
         ValueTask Collect(string path, CancellationToken token)
         {
             token.ThrowIfCancellationRequested();
@@ -27,22 +38,51 @@ partial class LocalClient
             return ValueTask.CompletedTask;
         }
 
-        var job0 = Parallel.ForEachAsync(Directory.EnumerateFiles(configSettings.OriginalFolder), token, Collect);
-        var job1 = Parallel.ForEachAsync(Directory.EnumerateFiles(configSettings.ThumbnailFolder), token, Collect);
+        var job0 = Parallel.ForEachAsync(Directory.EnumerateFiles(configSettings.OriginalFolder, "*", SearchOption.AllDirectories), token, Collect);
+        var job1 = Parallel.ForEachAsync(Directory.EnumerateFiles(configSettings.ThumbnailFolder, "*", SearchOption.AllDirectories), token, Collect);
         await job0.ConfigureAwait(false);
         await job1.ConfigureAwait(false);
 
-        logger.LogWarning($"{IOUtility.WarningColor}Are you sure to delete {files.Count} files? Input 'yes'.{IOUtility.NormalizeColor}");
-        var input = Console.ReadLine();
-        if (input == "yes")
+        if (!files.IsEmpty)
         {
-            await Parallel.ForEachAsync(files, token, (file, token) =>
+            logger.LogWarning($"{IOUtility.WarningColor}Are you sure to delete {files.Count} files? Input 'yes'.{IOUtility.NormalizeColor}");
+            var input = System.Console.ReadLine();
+            if (input == "yes")
             {
-                file.Delete();
-                return ValueTask.CompletedTask;
-            }).ConfigureAwait(false);
+                await Parallel.ForEachAsync(files, token, (file, token) =>
+                {
+                    file.Delete();
+                    return ValueTask.CompletedTask;
+                }).ConfigureAwait(false);
 
-            logger.LogWarning($"{IOUtility.WarningColor}Done.{IOUtility.NormalizeColor}");
+                logger.LogWarning($"{IOUtility.WarningColor}Done.{IOUtility.NormalizeColor}");
+            }
         }
+    }
+
+    private static async ValueTask ForceClearAsync(ILogger logger, ConfigSettings configSettings, bool pipe, CancellationToken token)
+    {
+        ulong clear = 0UL;
+        ValueTask Delete(string path, CancellationToken token)
+        {
+            token.ThrowIfCancellationRequested();
+            var info = new FileInfo(path);
+            if (info.Exists && info.Length == 0)
+            {
+                if (!pipe)
+                {
+                    logger.LogInformation(info.Name);
+                }
+
+                Interlocked.Increment(ref clear);
+                info.Delete();
+            }
+
+            return ValueTask.CompletedTask;
+        }
+
+        await Parallel.ForEachAsync(Directory.EnumerateFiles(configSettings.OriginalFolder, "*", SearchOption.AllDirectories), token, Delete).ConfigureAwait(false);
+        await Parallel.ForEachAsync(Directory.EnumerateFiles(configSettings.ThumbnailFolder, "*", SearchOption.AllDirectories), token, Delete).ConfigureAwait(false);
+        logger.LogInformation($"{IOUtility.WarningColor}{clear} files are deleted.{IOUtility.NormalizeColor}");
     }
 }
